{tc,xdp}: Change type of `data()` and `data_end()` to `*mut u8`
Returning the address of packet data as an integer brings a risk of LLVM
/ rustc performing bitwise operations on them (when, for example, a
program reads data from the packet with 2^x offsets), which is
(rightfully) prohibited by the eBPF verifier.

Returning a raw pointer is going to prevent this, because Rust will be
aware that it's a pointer, not an usual integer, and developers will
need to use `add()` or `offset()` methods (which are definitely not
going to be "optimized" as bitwise operations).

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>