aya-build: Allow to skip cargo-in-cargo builds
Not everyone likes cargo-in-cargo. The biggest advantage is that one
can't forget about rebuilding the eBPF crate. The biggest disadvantage
is that aya-build makes assumptions about which toolchain to use and
with which parameters.

It's also heavily designed for using with rustup and it's hard to use
for package maintainers who want to use custom Rust toolchains.

To appease people who prefer manual builds of eBPF crates, allow to
disable aya-build with `AYA_BUILD_SKIP`.

Fixes: #1329